add no_referral to nodes that are terminal and
on a path with no referrals. this fixes an issue
where a user has selected a special referral on
the homepage and then gets to a terminal node
with a tooltip that suggests the next page is to
cross-check.

in this rare case of a user with referrals on a
no referral path, we do not render the tooltip.